[MISC][WK] Rename file name to kebab case
